requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-calamine==0.2.3
rapidfuzz==3.10.0
scikit-learn==1.5.2
anthropic==0.39.0
python-multipart==0.0.12
python-dotenv==1.0.1
psycopg[binary]==3.2.3
psycopg-pool==3.2.3
orjson==3.10.11
